timer fixes

+++ devices/rp2/button_sequence_game.py
@@ -110,9 +110,9 @@ class ButtonSequenceGame:
         self.blink_led('white', 3, 0.5)
 
         # Add the button callback function to the GPIO event detection
-        GPIO.add_event_detect(self.BUTTON1, GPIO.FALLING, callback=self.button_callback, bouncetime=self.debounce_treshold*1000)
-        GPIO.add_event_detect(self.BUTTON2, GPIO.FALLING, callback=self.button_callback, bouncetime=self.debounce_treshold*1000)
-        GPIO.add_event_detect(self.BUTTON3, GPIO.FALLING, callback=self.button_callback, bouncetime=self.debounce_treshold*1000)
+        GPIO.add_event_detect(self.BUTTON1, GPIO.FALLING, callback=self.button_callback, bouncetime=int(self.debounce_treshold*1000))
+        GPIO.add_event_detect(self.BUTTON2, GPIO.FALLING, callback=self.button_callback, bouncetime=int(self.debounce_treshold*1000))
+        GPIO.add_event_detect(self.BUTTON3, GPIO.FALLING, callback=self.button_callback, bouncetime=int(self.debounce_treshold*1000))
 
         # Wait until the game is finished
         while not self.finished:
